+++ src/controller/controller.ts
@@ -18,6 +18,7 @@ import Database from "./database";
 import type * as Events from "./events";
 import GreenPower from "./greenPower";
 import {ZclFrameConverter} from "./helpers";
+import {checkInstallCode, parseInstallCode} from "./helpers/installCodes";
 import {Device, Entity} from "./model";
 import {InterviewState} from "./model/device";
 import Group from "./model/group";
@@ -246,34 +247,12 @@ export class Controller extends events.EventEmitter<ControllerEventMap> {
     }
 
     public async addInstallCode(installCode: string): Promise<void> {
-        const aqaraMatch = installCode.match(/^G\$M:.+\$A:(.+)\$I:(.+)$/);
-        const pipeMatch = installCode.match(/^(.+)\|(.+)$/);
-        let ieeeAddr: string;
-        let keyStr: string;
-
-        if (aqaraMatch) {
-            ieeeAddr = aqaraMatch[1];
-            keyStr = aqaraMatch[2];
-        } else if (pipeMatch) {
-            ieeeAddr = pipeMatch[1];
-            keyStr = pipeMatch[2];
-        } else {
-            assert(
-                installCode.length === 95 || installCode.length === 91,
-                `Unsupported install code, got ${installCode.length} chars, expected 95 or 91`,
-            );
-            const keyStart = installCode.length - (installCode.length === 95 ? 36 : 32);
-            ieeeAddr = installCode.substring(keyStart - 19, keyStart - 3);
-            keyStr = installCode.substring(keyStart, installCode.length);
-        }
-
-        ieeeAddr = `0x${ieeeAddr}`;
-        // match valid else asserted above
-        // biome-ignore lint/style/noNonNullAssertion: ignored using `--suppress`
+        // will throw if code cannot be parsed
+        const [ieeeAddr, keyStr] = parseInstallCode(installCode);
+        // biome-ignore lint/style/noNonNullAssertion: valid from above parsing
         const key = Buffer.from(keyStr.match(/.{1,2}/g)!.map((d) => Number.parseInt(d, 16)));
-
         // will throw if code cannot be fixed and is invalid
-        const [adjustedKey, adjusted] = ZSpec.Utils.checkInstallCode(key, true);
+        const [adjustedKey, adjusted] = checkInstallCode(key, true);
 
         if (adjusted) {
             logger.info(`Install code was adjusted for reason '${adjusted}'.`, NS);
@@ -383,7 +362,7 @@ export class Controller extends events.EventEmitter<ControllerEventMap> {
         if (this.options.backupPath && (await this.adapter.supportsBackup())) {
             logger.debug("Creating coordinator backup", NS);
             const backup = await this.adapter.backup(this.getDeviceIeeeAddresses());
-            const unifiedBackup = await BackupUtils.toUnifiedBackup(backup);
+            const unifiedBackup = BackupUtils.toUnifiedBackup(backup);
             const tmpBackupPath = `${this.options.backupPath}.tmp`;
             fs.writeFileSync(tmpBackupPath, JSON.stringify(unifiedBackup, null, 2));
             fs.renameSync(tmpBackupPath, this.options.backupPath);
@@ -657,7 +636,7 @@ export class Controller extends events.EventEmitter<ControllerEventMap> {
         this.emit("adapterDisconnected");
     }
 
-    private async onDeviceJoinedGreenPower(payload: GreenPowerDeviceJoinedPayload): Promise<void> {
+    private onDeviceJoinedGreenPower(payload: GreenPowerDeviceJoinedPayload): void {
         logger.debug(() => `Green power device '${JSON.stringify(payload).replaceAll(/\[[\d,]+\]/g, "HIDDEN")}' joined`, NS);
 
         // Green power devices don't have an ieeeAddr, the sourceID is unique and static so use this.
@@ -1001,13 +980,11 @@ export class Controller extends events.EventEmitter<ControllerEventMap> {
             }
 
             if (type === "readResponse" || type === "attributeReport") {
-                // Some device report, e.g. it's modelID through a readResponse or attributeReport
-                for (const key in data) {
-                    const property = Device.REPORTABLE_PROPERTIES_MAPPING[key];
-
-                    if (property && !device[property.key]) {
-                        // XXX: data technically can be `KeyValue | (string | number)[]`
-                        property.set((data as KeyValue)[key], device);
+                // devices report attributes through readRsp or attributeReport
+                if (frame.isCluster("genBasic")) {
+                    // data is `KeyValue` from type check above
+                    for (const key in data as KeyValue) {
+                        Device.REPORTABLE_PROPERTIES_MAPPING[key]?.set((data as KeyValue)[key], device);
                     }
                 }